requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
openai==1.6.1
streamlit==1.29.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.3
pyyaml==6.0.1
